pyproject.toml
[project]
name = "deeplabcut2yolo"
authors = [{ name = "Sira Pornsiriprasert", email = "code@psira.me" }]
maintainers = [{ name = "Sira Pornsiriprasert", email = "code@psira.me" }]
dynamic = ["version"]
description = "Convert DeepLabCut dataset to YOLO format"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = ["numpy", "pyyaml"]

[project.urls]
Homepage = "https://github.com/p-sira/deeplabcut2yolo"
Repository = "https://github.com/p-sira/deeplabcut2yolo"
Issues = "https://github.com/p-sira/deeplabcut2yolo/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "deeplabcut2yolo.__about__.__version__"}

[build-system]
requires = ["setuptools>=64"]